+++ src/main/java/bobo/erp/domain/rule/RuleLine.java
@@ -21,62 +21,62 @@ public class RuleLine {
     private Rule rule;
 
     private String line1Name;           //生产线名
-    private Double line1UnitInvest;     //单位投资
+    private Integer line1UnitInvest;     //单位投资
     private Integer line1InstallTime;   //安装时间
     private Integer line1ProduceTime;   //生产周期
-    private Double line1ChangeInvest;   //转产费用，每周期
+    private Integer line1ChangeInvest;   //转产费用，每周期
     private Integer line1ChangeTime;    //转产周期
-    private Double line1Upkeep;         //维护费
-    private Double line1ScrapValue;     //残值
-    private Double line1Depreciation;   //单位周期折旧
+    private Integer line1Upkeep;         //维护费
+    private Integer line1ScrapValue;     //残值
+    private Integer line1Depreciation;   //单位周期折旧
     private Integer line1DepreTime;     //折旧周期
     private Integer line1Score;         //潜力分数
 
     private String line2Name;
-    private Double line2UnitInvest;
+    private Integer line2UnitInvest;
     private Integer line2InstallTime;
     private Integer line2ProduceTime;
-    private Double line2ChangeInvest;
+    private Integer line2ChangeInvest;
     private Integer line2ChangeTime;
-    private Double line2Upkeep;
-    private Double line2ScrapValue;
-    private Double line2Depreciation;
+    private Integer line2Upkeep;
+    private Integer line2ScrapValue;
+    private Integer line2Depreciation;
     private Integer line2DepreTime;
     private Integer line2Score;
 
     private String line3Name;
-    private Double line3UnitInvest;
+    private Integer line3UnitInvest;
     private Integer line3InstallTime;
     private Integer line3ProduceTime;
-    private Double line3ChangeInvest;
+    private Integer line3ChangeInvest;
     private Integer line3ChangeTime;
-    private Double line3Upkeep;
-    private Double line3ScrapValue;
-    private Double line3Depreciation;
+    private Integer line3Upkeep;
+    private Integer line3ScrapValue;
+    private Integer line3Depreciation;
     private Integer line3DepreTime;
     private Integer line3Score;
 
     private String line4Name;
-    private Double line4UnitInvest;
+    private Integer line4UnitInvest;
     private Integer line4InstallTime;
     private Integer line4ProduceTime;
-    private Double line4ChangeInvest;
+    private Integer line4ChangeInvest;
     private Integer line4ChangeTime;
-    private Double line4Upkeep;
-    private Double line4ScrapValue;
-    private Double line4Depreciation;
+    private Integer line4Upkeep;
+    private Integer line4ScrapValue;
+    private Integer line4Depreciation;
     private Integer line4DepreTime;
     private Integer line4Score;
 
     private String line5Name;
-    private Double line5UnitInvest;
+    private Integer line5UnitInvest;
     private Integer line5InstallTime;
     private Integer line5ProduceTime;
-    private Double line5ChangeInvest;
+    private Integer line5ChangeInvest;
     private Integer line5ChangeTime;
-    private Double line5Upkeep;
-    private Double line5ScrapValue;
-    private Double line5Depreciation;
+    private Integer line5Upkeep;
+    private Integer line5ScrapValue;
+    private Integer line5Depreciation;
     private Integer line5DepreTime;
     private Integer line5Score;
 
@@ -106,11 +106,11 @@ public class RuleLine {
         this.line1Name = line1Name;
     }
 
-    public Double getLine1UnitInvest() {
+    public Integer getLine1UnitInvest() {
         return line1UnitInvest;
     }
 
-    public void setLine1UnitInvest(Double line1UnitInvest) {
+    public void setLine1UnitInvest(Integer line1UnitInvest) {
         this.line1UnitInvest = line1UnitInvest;
     }
 
@@ -130,11 +130,11 @@ public class RuleLine {
         this.line1ProduceTime = line1ProduceTime;
     }
 
-    public Double getLine1ChangeInvest() {
+    public Integer getLine1ChangeInvest() {
         return line1ChangeInvest;
     }
 
-    public void setLine1ChangeInvest(Double line1ChangeInvest) {
+    public void setLine1ChangeInvest(Integer line1ChangeInvest) {
         this.line1ChangeInvest = line1ChangeInvest;
     }
 
@@ -146,27 +146,27 @@ public class RuleLine {
         this.line1ChangeTime = line1ChangeTime;
     }
 
-    public Double getLine1Upkeep() {
+    public Integer getLine1Upkeep() {
         return line1Upkeep;
     }
 
-    public void setLine1Upkeep(Double line1Upkeep) {
+    public void setLine1Upkeep(Integer line1Upkeep) {
         this.line1Upkeep = line1Upkeep;
     }
 
-    public Double getLine1ScrapValue() {
+    public Integer getLine1ScrapValue() {
         return line1ScrapValue;
     }
 
-    public void setLine1ScrapValue(Double line1ScrapValue) {
+    public void setLine1ScrapValue(Integer line1ScrapValue) {
         this.line1ScrapValue = line1ScrapValue;
     }
 
-    public Double getLine1Depreciation() {
+    public Integer getLine1Depreciation() {
         return line1Depreciation;
     }
 
-    public void setLine1Depreciation(Double line1Depreciation) {
+    public void setLine1Depreciation(Integer line1Depreciation) {
         this.line1Depreciation = line1Depreciation;
     }
 
@@ -194,11 +194,11 @@ public class RuleLine {
         this.line2Name = line2Name;
     }
 
-    public Double getLine2UnitInvest() {
+    public Integer getLine2UnitInvest() {
         return line2UnitInvest;
     }
 
-    public void setLine2UnitInvest(Double line2UnitInvest) {
+    public void setLine2UnitInvest(Integer line2UnitInvest) {
         this.line2UnitInvest = line2UnitInvest;
     }
 
@@ -218,11 +218,11 @@ public class RuleLine {
         this.line2ProduceTime = line2ProduceTime;
     }
 
-    public Double getLine2ChangeInvest() {
+    public Integer getLine2ChangeInvest() {
         return line2ChangeInvest;
     }
 
-    public void setLine2ChangeInvest(Double line2ChangeInvest) {
+    public void setLine2ChangeInvest(Integer line2ChangeInvest) {
         this.line2ChangeInvest = line2ChangeInvest;
     }
 
@@ -234,27 +234,27 @@ public class RuleLine {
         this.line2ChangeTime = line2ChangeTime;
     }
 
-    public Double getLine2Upkeep() {
+    public Integer getLine2Upkeep() {
         return line2Upkeep;
     }
 
-    public void setLine2Upkeep(Double line2Upkeep) {
+    public void setLine2Upkeep(Integer line2Upkeep) {
         this.line2Upkeep = line2Upkeep;
     }
 
-    public Double getLine2ScrapValue() {
+    public Integer getLine2ScrapValue() {
         return line2ScrapValue;
     }
 
-    public void setLine2ScrapValue(Double line2ScrapValue) {
+    public void setLine2ScrapValue(Integer line2ScrapValue) {
         this.line2ScrapValue = line2ScrapValue;
     }
 
-    public Double getLine2Depreciation() {
+    public Integer getLine2Depreciation() {
         return line2Depreciation;
     }
 
-    public void setLine2Depreciation(Double line2Depreciation) {
+    public void setLine2Depreciation(Integer line2Depreciation) {
         this.line2Depreciation = line2Depreciation;
     }
 
@@ -282,11 +282,11 @@ public class RuleLine {
         this.line3Name = line3Name;
     }
 
-    public Double getLine3UnitInvest() {
+    public Integer getLine3UnitInvest() {
         return line3UnitInvest;
     }
 
-    public void setLine3UnitInvest(Double line3UnitInvest) {
+    public void setLine3UnitInvest(Integer line3UnitInvest) {
         this.line3UnitInvest = line3UnitInvest;
     }
 
@@ -306,11 +306,11 @@ public class RuleLine {
         this.line3ProduceTime = line3ProduceTime;
     }
 
-    public Double getLine3ChangeInvest() {
+    public Integer getLine3ChangeInvest() {
         return line3ChangeInvest;
     }
 
-    public void setLine3ChangeInvest(Double line3ChangeInvest) {
+    public void setLine3ChangeInvest(Integer line3ChangeInvest) {
         this.line3ChangeInvest = line3ChangeInvest;
     }
 
@@ -322,27 +322,27 @@ public class RuleLine {
         this.line3ChangeTime = line3ChangeTime;
     }
 
-    public Double getLine3Upkeep() {
+    public Integer getLine3Upkeep() {
         return line3Upkeep;
     }
 
-    public void setLine3Upkeep(Double line3Upkeep) {
+    public void setLine3Upkeep(Integer line3Upkeep) {
         this.line3Upkeep = line3Upkeep;
     }
 
-    public Double getLine3ScrapValue() {
+    public Integer getLine3ScrapValue() {
         return line3ScrapValue;
     }
 
-    public void setLine3ScrapValue(Double line3ScrapValue) {
+    public void setLine3ScrapValue(Integer line3ScrapValue) {
         this.line3ScrapValue = line3ScrapValue;
     }
 
-    public Double getLine3Depreciation() {
+    public Integer getLine3Depreciation() {
         return line3Depreciation;
     }
 
-    public void setLine3Depreciation(Double line3Depreciation) {
+    public void setLine3Depreciation(Integer line3Depreciation) {
         this.line3Depreciation = line3Depreciation;
     }
 
@@ -370,11 +370,11 @@ public class RuleLine {
         this.line4Name = line4Name;
     }
 
-    public Double getLine4UnitInvest() {
+    public Integer getLine4UnitInvest() {
         return line4UnitInvest;
     }
 
-    public void setLine4UnitInvest(Double line4UnitInvest) {
+    public void setLine4UnitInvest(Integer line4UnitInvest) {
         this.line4UnitInvest = line4UnitInvest;
     }
 
@@ -394,11 +394,11 @@ public class RuleLine {
         this.line4ProduceTime = line4ProduceTime;
     }
 
-    public Double getLine4ChangeInvest() {
+    public Integer getLine4ChangeInvest() {
         return line4ChangeInvest;
     }
 
-    public void setLine4ChangeInvest(Double line4ChangeInvest) {
+    public void setLine4ChangeInvest(Integer line4ChangeInvest) {
         this.line4ChangeInvest = line4ChangeInvest;
     }
 
@@ -410,27 +410,27 @@ public class RuleLine {
         this.line4ChangeTime = line4ChangeTime;
     }
 
-    public Double getLine4Upkeep() {
+    public Integer getLine4Upkeep() {
         return line4Upkeep;
     }
 
-    public void setLine4Upkeep(Double line4Upkeep) {
+    public void setLine4Upkeep(Integer line4Upkeep) {
         this.line4Upkeep = line4Upkeep;
     }
 
-    public Double getLine4ScrapValue() {
+    public Integer getLine4ScrapValue() {
         return line4ScrapValue;
     }
 
-    public void setLine4ScrapValue(Double line4ScrapValue) {
+    public void setLine4ScrapValue(Integer line4ScrapValue) {
         this.line4ScrapValue = line4ScrapValue;
     }
 
-    public Double getLine4Depreciation() {
+    public Integer getLine4Depreciation() {
         return line4Depreciation;
     }
 
-    public void setLine4Depreciation(Double line4Depreciation) {
+    public void setLine4Depreciation(Integer line4Depreciation) {
         this.line4Depreciation = line4Depreciation;
     }
 
@@ -458,11 +458,11 @@ public class RuleLine {
         this.line5Name = line5Name;
     }
 
-    public Double getLine5UnitInvest() {
+    public Integer getLine5UnitInvest() {
         return line5UnitInvest;
     }
 
-    public void setLine5UnitInvest(Double line5UnitInvest) {
+    public void setLine5UnitInvest(Integer line5UnitInvest) {
         this.line5UnitInvest = line5UnitInvest;
     }
 
@@ -482,11 +482,11 @@ public class RuleLine {
         this.line5ProduceTime = line5ProduceTime;
     }
 
-    public Double getLine5ChangeInvest() {
+    public Integer getLine5ChangeInvest() {
         return line5ChangeInvest;
     }
 
-    public void setLine5ChangeInvest(Double line5ChangeInvest) {
+    public void setLine5ChangeInvest(Integer line5ChangeInvest) {
         this.line5ChangeInvest = line5ChangeInvest;
     }
 
@@ -498,27 +498,27 @@ public class RuleLine {
         this.line5ChangeTime = line5ChangeTime;
     }
 
-    public Double getLine5Upkeep() {
+    public Integer getLine5Upkeep() {
         return line5Upkeep;
     }
 
-    public void setLine5Upkeep(Double line5Upkeep) {
+    public void setLine5Upkeep(Integer line5Upkeep) {
         this.line5Upkeep = line5Upkeep;
     }
 
-    public Double getLine5ScrapValue() {
+    public Integer getLine5ScrapValue() {
         return line5ScrapValue;
     }
 
-    public void setLine5ScrapValue(Double line5ScrapValue) {
+    public void setLine5ScrapValue(Integer line5ScrapValue) {
         this.line5ScrapValue = line5ScrapValue;
     }
 
-    public Double getLine5Depreciation() {
+    public Integer getLine5Depreciation() {
         return line5Depreciation;
     }
 
-    public void setLine5Depreciation(Double line5Depreciation) {
+    public void setLine5Depreciation(Integer line5Depreciation) {
         this.line5Depreciation = line5Depreciation;
     }
